requirements.txt
django>=4.2,<5.0
djangorestframework>=3.14,<4.0
django-cors-headers>=4.0,<5.0
python-dotenv>=1.0,<2.0
sentence-transformers>=2.2.0
numpy
jinja2
faiss-cpu
requests
PyPDF2>=3.0.0
python-docx>=1.0.0
Pillow>=10.0.0

